deps/pypros/requirements.txt
varint
GitPython